+++ src/main/java/project/closet/domain/users/AuthController.java
@@ -1,4 +1,4 @@
-package project.closet.domain.users.auth;
+package project.closet.domain.users;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -8,8 +8,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import project.closet.domain.users.auth.dto.ChangePasswordRequest;
-import project.closet.domain.users.auth.dto.SignInRequest;
+import project.closet.domain.users.dto.ChangePasswordRequest;
+import project.closet.domain.users.dto.SignInRequest;
 
 @RestController
 @RequestMapping("/api/auth")
