+++ src/main/java/project/closet/domain/users/dto/ProfileUpdateRequest.java
@@ -1,10 +1,10 @@
-package project.closet.domain.users.user.dto;
+package project.closet.domain.users.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import java.time.LocalDate;
 import project.closet.common.dto.Location;
-import project.closet.domain.users.user.Profile.TemperatureSensitivity;
-import project.closet.domain.users.user.User.Gender;
+import project.closet.domain.users.Profile.TemperatureSensitivity;
+import project.closet.domain.users.User.Gender;
 
 public record ProfileUpdateRequest(
     @NotBlank(message = "이름을 입력해주세요.")
