+++ src/main/java/project/closet/domain/users/dto/SignInRequest.java
@@ -1,4 +1,4 @@
-package project.closet.domain.users.auth.dto;
+package project.closet.domain.users.dto;
 
 public record SignInRequest(
     String email,
