+++ src/main/java/project/closet/domain/users/dto/UserLockUpdateRequest.java
@@ -1,4 +1,4 @@
-package project.closet.domain.users.user.dto;
+package project.closet.domain.users.dto;
 
 public record UserLockUpdateRequest(
     boolean locked) {
