+++ src/main/java/project/closet/domain/users/dto/UserRoleUpdateRequest.java
@@ -1,7 +1,7 @@
-package project.closet.domain.users.user.dto;
+package project.closet.domain.users.dto;
 
 import jakarta.validation.constraints.NotBlank;
-import project.closet.domain.users.user.User.UserRole;
+import project.closet.domain.users.User.UserRole;
 
 public record UserRoleUpdateRequest(
    @NotBlank(message = "역할을 입력해주세요.")
