+++ src/main/java/project/closet/domain/users/repository/UserRepository.java
@@ -1,8 +1,9 @@
-package project.closet.domain.users.user;
+package project.closet.domain.users.repository;
 
 import java.util.Optional;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
+import project.closet.domain.users.User;
 
 public interface UserRepository extends JpaRepository<User, UUID> {
