+++ src/main/java/project/closet/global/config/security/ClosetUserDetailService.java
@@ -7,9 +7,9 @@ import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
-import project.closet.domain.users.user.User;
-import project.closet.domain.users.user.User.UserRole;
-import project.closet.domain.users.user.UserRepository;
+import project.closet.domain.users.User;
+import project.closet.domain.users.User.UserRole;
+import project.closet.domain.users.repository.UserRepository;
 
 @RequiredArgsConstructor
 @Service
